Refactor TagUtils to Tagging Utils Fixup double-quote to left double-quote/rigght double-quote Fixup tests
