+++ src/test/resources/fl.json
@@ -23,7 +23,7 @@
     },
     "description": [
       "Supporting Act: HOTTUB",
-      "This item was digitized as part of the \"Cleared to Land\" project, supported by a grant from the National Historical Publications & Records Commission (NHPRC)."
+      "This item was digitized as part of the “Cleared to Land” project, supported by a grant from the National Historical Publications & Records Commission (NHPRC)."
     ],
     "extent": [
       "11 x 17 in."
